+++ cmd/atb/main.go
@@ -1,10 +1,12 @@
 // Package main provides the atb command entrypoint.
 package main
 
-import "fmt"
+import "os"
 
 var version = "dev"
 
 func main() {
-	fmt.Println(version)
+	if err := Execute(); err != nil {
+		os.Exit(1)
+	}
 }